fix:MiniLink repo split & add submodule
