+++ .gitmodules
@@ -1,0 +1,6 @@
+[submodule "MiniLink"]
+	path = MiniLink
+	url = https://github.com/NARAE-INHA-UNIV/MiniLink
+[submodule "backend/MiniLink"]
+	path = backend/MiniLink
+	url = https://github.com/NARAE-INHA-UNIV/MiniLink
