+++ backend/attitude_overview_manger.py
@@ -1,7 +1,7 @@
 import time
 
 from PySide6.QtCore import QObject, Signal, Slot
-from .lib.xmlHandler import XmlHandler
+from .MiniLink.lib.xmlHandler import XmlHandler
 
 
 class AttitudeOverviewManager(QObject):
