--- backend/lib/MSG/standard.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<mavlink>
-  <!-- MAVLink standard messages -->
-  <include>minimal.xml</include>
-  <dialect>0</dialect>
-  <enums>
-    <enum name="BOOL" bitmask="true">
-      <description>Enum used to indicate true or false (also: success or failure, enabled or disabled, active or inactive).</description>
-      <entry value="0" name="BOOL_FALSE">
-        <description>False.</description>
-      </entry>
-      <entry value="1" name="BOOL_TRUE">
-        <description>True.</description>
-      </entry>
-    </enum>
-  </enums>
-  <!-- use minimal.xml messages -->
-  <messages/>
-</mavlink>
